[MIPS] Replace disk_preparing stuff with setup_fs mechanism.
The setup_fs mechanism provides the same facility as the
disk_preparing code that we implemented but additionally
reboots the system starting the init sequence from scratch.

Removing config/init.rc file, it's not needed and likely
confusing to leave it in.

Not required for Android-4.3 but gets us more in sync with
other devices.

Signed-off-by: Pete Delaney <piet.delaney@imgtec.com>